Refactor and add initial swap.cow.fi landing event
