+++ src/custom/components/analytics/events/otherEvents.ts
@@ -1,5 +1,6 @@
-import { Category, sendEvent } from '../index'
+import { sendEvent } from '../index'
 import { detectExplorer } from 'utils/explorer'
+import { Category } from '../types'
 
 type GameType = 'CoW Runner' | 'MEV Slicer'
 export function gameAnalytics(gameType: GameType) {
@@ -19,4 +20,23 @@ export function externalLinkAnalytics(href: string) {
       label: href,
     })
   }
+}
+
+export function serviceWorkerAnalytics() {
+  const installed = Boolean(window.navigator.serviceWorker?.controller)
+  const hit = Boolean((window as any).__isDocumentCached)
+  const action = installed ? (hit ? 'Cache hit' : 'Cache miss') : 'Not installed'
+
+  sendEvent({
+    category: Category.SERVICE_WORKER,
+    nonInteraction: true,
+    action,
+  })
+}
+
+export function initAnalytics() {
+  sendEvent({
+    category: Category.INIT,
+    action: 'Initial landing on swap.cow.fi',
+  })
 }
