+++ src/custom/components/analytics/events/themeEvents.ts
@@ -1,4 +1,5 @@
-import { Category, sendEvent } from '../index'
+import { sendEvent } from '../index'
+import { Category } from '../types'
 
 export function toggleDarkModeAnalytics(darkMode: boolean) {
   sendEvent({
