+++ src/custom/components/analytics/hooks/useAnalyticsReporter.ts
@@ -8,7 +8,7 @@ import { useWalletDetails, useWalletInfo } from '@cow/modules/wallet'
 import { getConnectionName, getIsMetaMask } from '@cow/modules/wallet/api/utils/connection'
 import { getWeb3ReactConnection } from '@cow/modules/wallet/web3-react/connection'
 import { googleAnalytics, GOOGLE_ANALYTICS_CLIENT_ID_STORAGE_KEY } from '..'
-import { Dimensions } from '../GoogleAnalyticsProvider'
+import { Dimensions } from '../types'
 import usePrevious from 'hooks/usePrevious'
 
 import { PIXEL_EVENTS } from '../pixel/constants'
